Skip encoding qualifiers before conversion
Type encoding like `const char*` would have a `r` in front of the
`*`. An example would be `NSString#stringWithCString`. Skipping all
type qualifiers before we head into conversion.

Fixing this resulted in exposing a bug where the NSData was being
freed on dealloc but we don't want that. May want to consider in the
future if we even need to do this NSMutableData part.