Convert Hash/Array to NSDictionary/NSArray
Use the dictionaryWithRubyHash and arrayWithRubyArray extension
methods internally when encountering a T_HASH or T_ARRAY within
RIGSCore. This way, we don't need to use the externally as you can
always just use dictionaryWithDictionary and arrayWithArray
externally, passing in a Ruby Hash or Array, and we will convert them
for you.