Adjust `objc_msgSend` for structs on Intel (#6)
While there are some ABI  between Intel and ARM on when exactly to
use `objc_msgSend_stret` and `objc_msgSend_fpret` but it brings
us a bit closer to reality.

Additionally, we need to modify our need to always expose clases by
their public name, as some classes like `NSAttributedString` have
`NSConcrete` versions that are returned in order to implement certain
methods. We see this in `NSAttributedString#initWithString` For now,
we check the internal class for `NSConcrete*` and allow that to go
through.