Confirm Ruby to ObjC selectors match args
Now that we auto-proxy subclasses in Ruby, we should be more careful
about what all methods we proxy. We have a contract between the two
languages where a ObjC selector's ":" becomes an methods "_" in Ruby.

But a class in Ruby could have regular snake case methods that don't
match our expectations based on the number of arguments. In cases like
these, we should skip proxying these methods, as a call from ObjC
would fail to pass the correct number of arguments.

Closes #14