Use BridgeSupport for protocols
The trouble with objc_getProtocol is that it doesn't work unless you
are explicitly referencing that protocol. Sometimes something inside
of Objective-C is doing this for us, but oftentimes not. So we added a
bunch of random references to protocols on load of this extension just
to bring them in for us as a work around.

But BridgeSupport provides many of these, and now that we can extend
BridgeSupport with our own custom support files, I feel better about
relying on it. So this makes the change to load known protocol
selectors from there instead of asking the runtime.

Fixes #19