Improve handling of structs (#32)
We have one spot where we need to build the FFI type for a struct and
we don't need to know the name of it, just its arity and its
individual types.

This adds a utility function in order to check on that and uses it
instead of checking knownStructs.